Update decks page styling + add loader/spinner

+++ frontend/src/components/Decks/index.tsx
@@ -2,7 +2,21 @@ import { useContext, useEffect, useState } from "react";
 import { useNavigate } from "react-router";
 import AuthContext from "@contexts/AuthContext";
 import { deleteDeckOfCards, getDecks } from "@lib/decks";
-import type { Deck } from "@utils/types";
+import type { Deck, DeckNoUserId } from "@utils/types";
+import {
+  Button,
+  Card,
+  CardInfo,
+  CardDelete,
+  CardTitle,
+  Grid,
+  Header,
+  HeaderOne,
+  Wrapper,
+  EmptyState,
+  EmptyText,
+} from "@globalStyles";
+import Loader from "@components/Loader";
 
 const Decks = () => {
   const auth = useContext(AuthContext);
@@ -29,34 +43,70 @@ const Decks = () => {
     fetchDecks();
   }, [auth?.session]);
 
-  const handleDelete = async (id: number) => {
-    await deleteDeckOfCards(id, token);
-    setDecks(decks.filter((deck) => deck.id !== Number(id)));
+  const handleEdit = async (
+    e: React.MouseEvent,
+    deckId: DeckNoUserId["id"]
+  ) => {
+    e.stopPropagation();
+
+    navigate(`/decks/${deckId}?action=edit`);
   };
 
-  if (loading) return <p>Loading</p>;
+  const handleDelete = async (
+    e: React.MouseEvent,
+    id: DeckNoUserId["id"],
+    title: DeckNoUserId["title"]
+  ) => {
+    e.stopPropagation();
+    const confirmed = confirm(
+      `Are you sure you'd like to delete the '${title}' deck?`
+    );
+    if (confirmed) {
+      await deleteDeckOfCards(id, token);
+      setDecks(decks.filter((deck) => deck.id !== Number(id)));
+    }
+  };
+
+  if (loading) return <Loader />;
 
   return (
-    <>
-      <h1>Decks</h1>
-      <button onClick={() => navigate("/decks/new")}>Create New Deck</button>
+    <Wrapper>
+      <Header>
+        <HeaderOne>My Decks</HeaderOne>
+        <Button onClick={() => navigate("/decks/new")}>Create New Deck</Button>
+      </Header>
       {decks?.length > 0 ? (
-        decks.map((deck) => {
-          const { title, id } = deck;
-          return (
-            <div key={id}>
-              <p onClick={() => navigate(`${id}`)}>{title}</p>
-              <i
-                className="fa-solid fa-trash"
-                onClick={() => handleDelete(id)}
-              ></i>
-            </div>
-          );
-        })
+        <Grid>
+          {decks.map((deck) => {
+            const { title, id, totalCards } = deck;
+            return (
+              <Card key={id} onClick={() => navigate(`${id}`)}>
+                <CardDelete
+                  onClick={(e) => handleDelete(e, id, title)}
+                  title="Delete Deck"
+                >
+                  <i className="fa-solid fa-trash"></i>
+                </CardDelete>
+                <CardTitle>{title}</CardTitle>
+                <CardInfo>
+                  <p>
+                    {totalCards} {totalCards === 1 ? "Card" : "Cards"}
+                  </p>
+                  <i
+                    className="fa-solid fa-pen-to-square"
+                    onClick={(e) => handleEdit(e, id)}
+                  ></i>
+                </CardInfo>
+              </Card>
+            );
+          })}
+        </Grid>
       ) : (
-        <p>You don't have any decks yet :(</p>
+        <EmptyState>
+          <EmptyText>You don't have any decks yet :(</EmptyText>
+        </EmptyState>
       )}
-    </>
+    </Wrapper>
   );
 };
